+++ app/src/main/java/com/internetEnemies/combatCritters/data/OffersDatabase.java
@@ -15,6 +15,7 @@ import com.internetEnemies.combatCritters.data.hsqldb.MarketRegistryHSQLDB;
 import com.internetEnemies.combatCritters.data.hsqldb.RewardTransactionRegistryHSQLDB;
 import com.internetEnemies.combatCritters.data.hsqldb.TradeTransactionRegistryHSQLDB;
 import com.internetEnemies.combatCritters.data.hsqldb.TransactionRegistryHSQLDB;
+import com.internetEnemies.combatCritters.data.market.IMarketDB;
 import com.internetEnemies.combatCritters.objects.TradeTransaction;
 import com.internetEnemies.combatCritters.objects.battles.RewardTransaction;
